+++ Side-dish/Side-dish/Main.storyboard
@@ -9,10 +9,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--View Controller-->
+        <!--Intial View Controller-->
         <scene sceneID="C9u-pu-Ofs">
             <objects>
-                <viewController storyboardIdentifier="InitialViewController" id="HW5-If-6Pr" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="InitialViewController" id="HW5-If-6Pr" customClass="IntialViewController" customModule="Side_dish" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="a1V-uc-aew">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -476,7 +476,7 @@
                                                 <rect key="frame" x="16" y="355" width="382" height="96.5"/>
                                                 <subviews>
                                                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Title" textAlignment="natural" lineBreakMode="characterWrap" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Btu-06-MM9">
-                                                        <rect key="frame" x="0.0" y="0.0" width="50" height="29"/>
+                                                        <rect key="frame" x="0.0" y="0.0" width="49" height="29"/>
                                                         <fontDescription key="fontDescription" type="boldSystem" pointSize="24"/>
                                                         <nil key="textColor"/>
                                                         <nil key="highlightedColor"/>
